+++ R/shoji.R
@@ -87,6 +87,39 @@ setEntitySlot <- function(x, i, value) {
     return(x)
 }
 
+#' Setter for Crunch objects that wraps in a "body"
+#'
+#' Variable Folders require an extra list(body = ...)
+#' compared to other places where we use `setEntitySlot()`
+#' is used. This may be a bug, see pivotal:
+#' <https://www.pivotaltracker.com/n/projects/2172644/stories/174429283>
+#'
+#' @param x a ShojiObject or subclass thereof
+#' @param i character the slot name to update
+#' @param value whatever the new value of that slot should be
+#' @return x modified accordingly.
+#' @keywords internal
+setEntitySlotWrapBody <- function(x, i, value) {
+    ## Check if we have actual changes to send. Wrap both sides in I()
+    ## in case "value" is already wrapped
+    if (!identical(I(slot(x, "body")[[i]]), I(value))) {
+        slot(x, "body")[[i]] <- value
+        body <- wrapCatalog(body = structure(list(value), .Names = i))
+        payload <- toJSON(body)
+        crPATCH(self(x), body = payload)
+        if (is.dataset(x)) {
+            # Update the tuple in place too
+            # This is hacky; we should probably make datasets not involve tuples
+            if (i %in% names(tuple(x)@body)) {
+                tuple(x)[[i]] <- value
+            }
+            # Also drop cache for the dataset's containing project index
+            dropOnly(shojiURL(x, "catalogs", "project"))
+        }
+    }
+    return(x)
+}
+
 #' Get a resource URL from a Shoji Object
 #'
 #' @param x a shojiObject